fix(fcm): Exposed admin.messaging namespace
