+++ src/firebase-namespace.d.ts
@@ -16,4 +16,5 @@
 
 export * from './credential/index';
 export * from './firebase-namespace-api';
+export * from './messaging/index';
 export * from './remote-config/index';
