+++ src/messaging/messaging-api-request-internal.ts
@@ -20,8 +20,11 @@ import {
 } from '../utils/api-request';
 import { createFirebaseError, getErrorCode } from './messaging-errors-internal';
 import { SubRequest, BatchRequestClient } from './batch-request-internal';
-import { SendResponse, BatchResponse } from './messaging-types';
+import { messaging } from './index';
 import { getSdkVersion } from '../utils/index';
+
+import SendResponse = messaging.SendResponse;
+import BatchResponse = messaging.BatchResponse;
 
 // FCM backend constants
 const FIREBASE_MESSAGING_TIMEOUT = 10000;
