+++ src/messaging/messaging.ts
@@ -18,20 +18,28 @@ import { FirebaseApp } from '../firebase-app';
 import { deepCopy, deepExtend } from '../utils/deep-copy';
 import { SubRequest } from './batch-request-internal';
 import { validateMessage, BLACKLISTED_DATA_PAYLOAD_KEYS, BLACKLISTED_OPTIONS_KEYS } from './messaging-internal';
-import {
-  Message, MessagingDevicesResponse,
-  MessagingDeviceGroupResponse, MessagingTopicManagementResponse,
-  MessagingPayload, MessagingOptions, MessagingTopicResponse,
-  MessagingConditionResponse, BatchResponse, MulticastMessage, DataMessagePayload, NotificationMessagePayload,
-} from './messaging-types';
+import { messaging } from './index';
 import { FirebaseMessagingRequestHandler } from './messaging-api-request-internal';
 import { FirebaseServiceInterface, FirebaseServiceInternalsInterface } from '../firebase-service';
-import {
-  ErrorInfo, MessagingClientErrorCode, FirebaseMessagingError,
-} from '../utils/error';
-
+import { ErrorInfo, MessagingClientErrorCode, FirebaseMessagingError } from '../utils/error';
 import * as utils from '../utils';
 import * as validator from '../utils/validator';
+
+import MessagingInterface = messaging.Messaging;
+import Message = messaging.Message;
+import BatchResponse = messaging.BatchResponse;
+import MulticastMessage = messaging.MulticastMessage;
+import MessagingTopicManagementResponse = messaging.MessagingTopicManagementResponse;
+
+// Legacy API types
+import MessagingDevicesResponse = messaging.MessagingDevicesResponse;
+import MessagingDeviceGroupResponse = messaging.MessagingDeviceGroupResponse;
+import MessagingPayload = messaging.MessagingPayload;
+import MessagingOptions = messaging.MessagingOptions;
+import MessagingTopicResponse = messaging.MessagingTopicResponse;
+import MessagingConditionResponse = messaging.MessagingConditionResponse;
+import DataMessagePayload = messaging.DataMessagePayload;
+import NotificationMessagePayload = messaging.NotificationMessagePayload;
 
 /* eslint-disable @typescript-eslint/camelcase */
 
@@ -196,7 +204,7 @@ class MessagingInternals implements FirebaseServiceInternalsInterface {
 /**
  * Messaging service bound to the provided app.
  */
-export class Messaging implements FirebaseServiceInterface {
+export class Messaging implements FirebaseServiceInterface, MessagingInterface {
 
   public INTERNAL: MessagingInternals = new MessagingInternals();
 
@@ -399,7 +407,7 @@ export class Messaging implements FirebaseServiceInterface {
     registrationTokenOrTokens: string | string[],
     payload: MessagingPayload,
     options: MessagingOptions = {},
-  ): Promise<MessagingDevicesResponse | MessagingDeviceGroupResponse> {
+  ): Promise<MessagingDevicesResponse> {
     // Validate the input argument types. Since these are common developer errors when getting
     // started, throw an error instead of returning a rejected promise.
     this.validateRegistrationTokensType(
@@ -437,7 +445,13 @@ export class Messaging implements FirebaseServiceInterface {
         if ('multicast_id' in response) {
           return mapRawResponseToDevicesResponse(response);
         } else {
-          return mapRawResponseToDeviceGroupResponse(response);
+          const groupResponse = mapRawResponseToDeviceGroupResponse(response);
+          return {
+            ...groupResponse,
+            canonicalRegistrationTokenCount: -1,
+            multicastId: -1,
+            results: [],
+          }
         }
       });
   }
@@ -463,7 +477,7 @@ export class Messaging implements FirebaseServiceInterface {
     notificationKey: string,
     payload: MessagingPayload,
     options: MessagingOptions = {},
-  ): Promise<MessagingDeviceGroupResponse | MessagingDevicesResponse> {
+  ): Promise<MessagingDeviceGroupResponse> {
     if (!validator.isNonEmptyString(notificationKey)) {
       throw new FirebaseMessagingError(
         MessagingClientErrorCode.INVALID_RECIPIENT,
@@ -515,7 +529,11 @@ export class Messaging implements FirebaseServiceInterface {
               'Notification key provided to sendToDeviceGroup() is invalid.',
             );
           } else {
-            return mapRawResponseToDevicesResponse(response);
+            const devicesResponse = mapRawResponseToDevicesResponse(response);
+            return {
+              ...devicesResponse,
+              failedRegistrationTokens: [],
+            }
           }
         }
